Crate cumulus_client_cli

Cumulus CLI library.

Structs

CollatorOptions

Options only relevant for collator nodes

ExportGenesisStateCommand

Command for exporting the genesis state of the parachain

ExportGenesisWasmCommand

Command for exporting the genesis wasm file.

NormalizedRunCmd

A non-redundant version of the RunCmd that sets the validator field when the original RunCmd had the collator field. This is how we make --collator imply --validator.

PurgeChainCmd

The purge-chain command used to remove the whole chain: the parachain and the relay chain.

RunCmd

The run command used to run a node.

Enums

RelayChainMode

Possible modes for the relay chain to operate in.

Functions

extract_genesis_wasm

Extract the genesis code from a given ChainSpec.

generate_genesis_block

Generate the genesis block from a given ChainSpec.